llm/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice-assistant"
version = "0.1.0"
description = "Vosk-based voice-to-CLI automation tool with wake word detection"
requires-python = ">=3.9"
dependencies = [
    "vosk>=0.3.45",
    "sounddevice>=0.4.6",
    "pyyaml>=6.0",
    "rapidfuzz>=3.0.0",
    "numpy"
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "mbake",
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "pyinstaller>=6.0.0",
]

[project.scripts]
voice-assistant = "voice_assistant.__main__:main"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
